feat: better error reporting, eyre integration

+++ crates/github_scbot/src/shell/commands/pulls.rs
@@ -14,11 +14,8 @@ use super::errors::Result;
 pub(crate) fn show_pull_request(config: &Config, repository_path: &str, number: u64) -> Result<()> {
     let conn = establish_single_connection(config)?;
 
-    let pr = PullRequestModel::get_from_repository_path_and_number(
-        &conn,
-        &repository_path,
-        number as i32,
-    )?;
+    let pr =
+        PullRequestModel::get_from_repository_path_and_number(&conn, &repository_path, number)?;
     println!(
         "Accessing pull request #{} on repository {}",
         number, repository_path
